requirements.txt
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
blinker==1.9.0
cachetools==5.5.2
celery==5.3.4
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.36.0
colorama==0.4.6
cryptography==45.0.3
Django==4.2.10
django-cors-headers==4.3.1
djangorestframework==3.14.0
ecdsa==0.19.1
fastapi==0.115.12
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.8
httpx==0.25.2
idna==3.10
Jinja2==3.1.6
jose==1.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.2
MarkupSafe==3.0.2
narwhals==1.34.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
passlib==1.7.4
Pillow==10.1.0
prompt_toolkit==3.0.50
protobuf==5.29.4
pyarrow==19.0.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==5.0.1
referencing==0.36.2
requests==2.31.0
rpds-py==0.24.0
rsa==4.9.1
six==1.17.0
slugify==0.0.1
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
sqlparse==0.5.3
starlette==0.46.1
streamlit==1.44.1
tenacity==9.1.2
toml==0.10.2
tornado==6.4.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.0
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
whitenoise==6.5.0
